Flatten Submissions

This function will flatten submissions coming from a FormIO list that has nested elements

flatten_submission_records(x)

Arguments

x

list with responses, typically obtained through the fetch_form_responses() function.

Value

list with

FlatResponses which is the entire dataset obtained from fetch_form_responses() but, where all the nested lists are flattened into a bidimensional data frame. nested lists will take their name by using their parent name as a prefix, then a "-" separator, and then their own name.

ColumnNames which is a data frame with column Number and column Names, listing all the newly created column names in FlatResponses and their numerical order. This output can then be passed to other FormIO functions for making easier to subset the dataset

Examples

x<-FoodTypes

# Nested Structure
head(x)
#>   form.submission_id            form.created_at form.user late_entry
#> 1             food-1 2026-01-28 11:52:28.070558    user_A      FALSE
#> 2             food-2 2026-01-28 11:52:27.070558    user_B      FALSE
#> 3             food-3 2026-01-28 11:52:26.070558    user_C      FALSE
#> 4             food-4 2026-01-28 11:52:25.070558    user_D      FALSE
#> 5             food-5 2026-01-28 11:52:24.070558    user_E      FALSE
#> 6             food-6 2026-01-28 11:52:23.070558    user_F      FALSE
#>   food_category  region is_category_correct  food_name
#> 1         Fruit   North                 yes      Apple
#> 2     Vegetable   South                 yes     Carrot
#> 3          Meat    East                 yes     Salmon
#> 4         Grain    West                 yes       Rice
#> 5         Dairy Central                 yes     Cheese
#> 6         Mixed Coastal                 yes Mixed Bowl
#>               food_description   status completion_pct select_chef.label
#> 1        Fresh seasonal apples  planned             10            Chef A
#> 2 Root vegetable used in soups prepared             40            Chef B
#> 3           Wild-caught salmon   served             90            Chef C
#> 4         Staple grain product  planned             20            Chef A
#> 5      Fermented dairy product   served             80            Chef B
#> 6 Combination of various foods prepared             50            Chef C
#>   select_chef.value select_recipe.label select_recipe.value
#> 1            Chef A           Apple Pie           Apple Pie
#> 2            Chef B         Carrot Soup         Carrot Soup
#> 3            Chef C      Grilled Salmon      Grilled Salmon
#> 4            Chef A           Rice Bowl           Rice Bowl
#> 5            Chef B        Cheese Plate        Cheese Plate
#> 6            Chef C         Fusion Dish         Fusion Dish
#>             final_notes
#> 1 Seasonal availability
#> 2    Requires prep time
#> 3          High protein
#> 4           Bulk staple
#> 5          Aged product
#> 6            Custom mix
#>                                                       editGrid total_cost
#> 1                                                         NULL         NA
#> 2                   Flour, Sugar, 1, 2, 5, 10, Local, Imported         15
#> 3 Carrot, Onion, Garlic, 5, 2, 1, 3, 2, 1, Local, Local, Local          6
#> 4                                          Salmon, 2, 20, Wild         20
#> 5            Rice, Beans, 10, 5, 15, 8, staple, grain, protein         23
#> 6                                                         NULL          0
#>   total_items
#> 1           0
#> 2           2
#> 3           3
#> 4           1
#> 5           2
#> 6           0

#Flattened Structure
xFlat<-flatten_submission_records(x)

xFlat$FlatResponses ## Survey Output
#> # A tibble: 10 × 24
#>    `form-submission_id` `form-created_at`   `form-user` late_entry food_category
#>    <chr>                <chr>               <chr>       <lgl>      <chr>        
#>  1 food-1               2026-01-28 11:52:2… user_A      FALSE      Fruit        
#>  2 food-2               2026-01-28 11:52:2… user_B      FALSE      Vegetable    
#>  3 food-2               2026-01-28 11:52:2… user_B      FALSE      Vegetable    
#>  4 food-3               2026-01-28 11:52:2… user_C      FALSE      Meat         
#>  5 food-3               2026-01-28 11:52:2… user_C      FALSE      Meat         
#>  6 food-3               2026-01-28 11:52:2… user_C      FALSE      Meat         
#>  7 food-4               2026-01-28 11:52:2… user_D      FALSE      Grain        
#>  8 food-5               2026-01-28 11:52:2… user_E      FALSE      Dairy        
#>  9 food-5               2026-01-28 11:52:2… user_E      FALSE      Dairy        
#> 10 food-6               2026-01-28 11:52:2… user_F      FALSE      Mixed        
#> # ℹ 19 more variables: region <chr>, is_category_correct <chr>,
#> #   food_name <chr>, food_description <chr>, status <chr>,
#> #   completion_pct <dbl>, `select_chef-label` <chr>, `select_chef-value` <chr>,
#> #   `select_recipe-label` <chr>, `select_recipe-value` <chr>,
#> #   final_notes <chr>, editGrid <lgl>, total_cost <dbl>, total_items <dbl>,
#> #   `editGrid-ingredient` <chr>, `editGrid-quantity` <dbl>,
#> #   `editGrid-cost` <dbl>, `editGrid-source` <chr>, `editGrid-tags` <I<list>>
xFlat$ColumnNames   ## New Columns Formed
#>    Number                Name
#> 1       1  form-submission_id
#> 2       2     form-created_at
#> 3       3           form-user
#> 4       4          late_entry
#> 5       5       food_category
#> 6       6              region
#> 7       7 is_category_correct
#> 8       8           food_name
#> 9       9    food_description
#> 10     10              status
#> 11     11      completion_pct
#> 12     12   select_chef-label
#> 13     13   select_chef-value
#> 14     14 select_recipe-label
#> 15     15 select_recipe-value
#> 16     16         final_notes
#> 17     17            editGrid
#> 18     18          total_cost
#> 19     19         total_items
#> 20     20 editGrid-ingredient
#> 21     21   editGrid-quantity
#> 22     22       editGrid-cost
#> 23     23     editGrid-source
#> 24     24       editGrid-tags